refactor: move router into resolver crate
