+++ crates/resolvers/src/lib.rs
@@ -1,10 +1,14 @@
+use application_utils::AuthContext;
 use async_graphql::{EmptySubscription, MergedObject, Schema};
+use async_graphql_axum::{GraphQLRequest, GraphQLResponse};
+use axum::Extension;
 
 pub use exporter_resolver::*;
 pub use file_storage_resolver::*;
 pub use fitness_resolver::*;
 pub use importer_resolver::*;
 pub use miscellaneous_resolver::*;
+pub use router_resolver::*;
 
 #[derive(MergedObject, Default)]
 pub struct QueryRoot(
@@ -25,3 +29,11 @@ pub struct MutationRoot(
 );
 
 pub type GraphqlSchema = Schema<QueryRoot, MutationRoot, EmptySubscription>;
+
+pub async fn graphql_handler(
+    schema: Extension<GraphqlSchema>,
+    gql_ctx: AuthContext,
+    req: GraphQLRequest,
+) -> GraphQLResponse {
+    schema.execute(req.into_inner().data(gql_ctx)).await.into()
+}
